Extract everything related to pattern deconstruction to a new module
